test: replace test cases for UserIdModel to model/foundation directory
